[#100] Remove Notify trait
