+++ iceoryx2/src/prelude.rs
@@ -14,7 +14,7 @@ pub use crate::iox2::Iox2;
 pub use crate::iox2::Iox2Event;
 pub use crate::payload_mut::{PayloadMut, UninitPayloadMut};
 pub use crate::port::{
-    event_id::EventId, notify::Notify, publish::DefaultLoan, publish::Publish, publish::SendCopy,
+    event_id::EventId, publish::DefaultLoan, publish::Publish, publish::SendCopy,
     publish::UninitLoan,
 };
 pub use crate::service::{process_local, service_name::ServiceName, zero_copy, Service};
