+++ iceoryx2/src/service/port_factory/notifier.rs
@@ -30,7 +30,7 @@
 //! ```
 use std::fmt::Debug;
 
-use crate::port::{event_id::EventId, notifier::Notifier, notify::NotifierCreateError};
+use crate::port::{event_id::EventId, notifier::Notifier, notifier::NotifierCreateError};
 use iceoryx2_bb_log::fail;
 
 use crate::service;
